server/requirements.txt
# Production dependencies for Resume Screening Environment
openenv-core[core]>=0.2.2
fastapi>=0.115.0
uvicorn[standard]>=0.24.0